requirements.txt
openai>=1.0.0
faiss-cpu>=1.7.4
pandas>=1.3.0
numpy>=1.20.0
python-dotenv>=0.19.0
tqdm>=4.62.0
flask>=2.0.0
flask-cors>=3.0.10
python-dateutil>=2.8.2
markdown>=3.3.0
